requirements.txt
django==4.2.3
djangorestframework==3.14.0
psycopg2-binary==2.9.6
requests==2.31.0
drf-yasg==1.21.6

requests~=2.31.0
rest_framework_simplejwt==0.0.2
django-filter==23.2
